+++ CHANGELOG.md
@@ -11,6 +11,7 @@ CHANGED:
 
 - update of model-object version to 1.8
 - update of model-type-safe version to 1.8
+- extracted project properties to gradle.properties
 
 Version 5.2
 -------------